requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai==1.65.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-telegram-bot==21.11.1
pytz==2025.2
requests==2.32.4
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.5.0
uvicorn==0.34.2
Levenshtein==0.27.1
python-Levenshtein==0.27.1
RapidFuzz==3.13.0
schedule==1.2.2
thefuzz==0.22.1
beautifulsoup4==4.13.4
sgmllib3k==1.0.0
soupsieve==2.7
feedparser==6.0.11
